+++ lib/src/states/camera_picker_state.dart
@@ -57,8 +57,8 @@ class CameraPickerState extends State<CameraPicker>
   /// 可用的相机实例
   late List<CameraDescription> cameras;
 
-  /// Whether the controller is handling taking picture or recording video.
-  /// 相机控制器是否在处理拍照或录像
+  /// Whether the controller is handling method calls.
+  /// 相机控制器是否在处理方法调用
   bool isControllerBusy = false;
 
   /// Current exposure offset.
@@ -186,6 +186,12 @@ class CameraPickerState extends State<CameraPicker>
     return pickerConfig.minimumRecordingDuration;
   }
 
+  /// Whether the capture button is displaying.
+  bool get shouldCaptureButtonDisplay =>
+      isControllerBusy ||
+      (innerController?.value.isRecordingVideo ?? false) &&
+          isRecordingRestricted;
+
   /// Whether the camera preview should be rotated.
   bool get isCameraRotated => pickerConfig.cameraQuarterTurns % 4 != 0;
 
@@ -258,6 +264,7 @@ class CameraPickerState extends State<CameraPicker>
     } else if (state == AppLifecycleState.inactive) {
       c.dispose();
       innerController = null;
+      isControllerBusy = false;
     }
   }
 
@@ -265,8 +272,11 @@ class CameraPickerState extends State<CameraPicker>
   /// 根据 [constraints] 获取相机预览适用的缩放。
   double effectiveCameraScale(
     BoxConstraints constraints,
-    CameraController controller,
+    CameraController? controller,
   ) {
+    if (controller == null) {
+      return 1;
+    }
     final int turns = cameraQuarterTurns;
     final String orientation = controller.value.deviceOrientation.toString();
     // Fetch the biggest size from the constraints.
@@ -831,7 +841,10 @@ class CameraPickerState extends State<CameraPicker>
     if (isControllerBusy) {
       return;
     }
-    isControllerBusy = true;
+    setState(() {
+      isControllerBusy = true;
+      isShootingButtonAnimate = true;
+    });
     final ExposureMode previousExposureMode = controller.value.exposureMode;
     try {
       await Future.wait(<Future<void>>[
@@ -881,8 +894,10 @@ class CameraPickerState extends State<CameraPicker>
     } catch (e, s) {
       handleErrorWithHandler(e, s, pickerConfig.onError);
     } finally {
-      isControllerBusy = false;
-      safeSetState(() {});
+      safeSetState(() {
+        isControllerBusy = false;
+        isShootingButtonAnimate = false;
+      });
     }
   }
 
@@ -909,14 +924,7 @@ class CameraPickerState extends State<CameraPicker>
   /// 将被取消，并且状态会重置。
   void recordDetectionCancel(PointerUpEvent event) {
     recordDetectTimer?.cancel();
-    if (isShootingButtonAnimate) {
-      safeSetState(() {
-        isShootingButtonAnimate = false;
-      });
-    }
     if (innerController?.value.isRecordingVideo == true) {
-      lastShootingButtonPressedPosition = null;
-      safeSetState(() {});
       stopRecordingVideo();
     }
   }
@@ -940,7 +948,6 @@ class CameraPickerState extends State<CameraPicker>
         ..reset()
         ..start();
     } catch (e, s) {
-      isControllerBusy = false;
       if (!controller.value.isRecordingVideo) {
         handleErrorWithHandler(e, s, pickerConfig.onError);
         return;
@@ -955,26 +962,31 @@ class CameraPickerState extends State<CameraPicker>
         recordStopwatch.stop();
       }
     } finally {
-      safeSetState(() {});
+      safeSetState(() {
+        isControllerBusy = false;
+      });
     }
   }
 
   /// Stop the recording process.
   /// 停止录制视频
   Future<void> stopRecordingVideo() async {
-    void handleError() {
-      recordCountdownTimer?.cancel();
-      isShootingButtonAnimate = false;
-      safeSetState(() {});
+    if (isControllerBusy) {
+      return;
     }
 
     recordStopwatch.stop();
-    if (!controller.value.isRecordingVideo) {
-      handleError();
+    if (innerController == null || !controller.value.isRecordingVideo) {
+      recordCountdownTimer?.cancel();
+      safeSetState(() {
+        isControllerBusy = false;
+        isShootingButtonAnimate = false;
+      });
       return;
     }
     safeSetState(() {
-      isShootingButtonAnimate = false;
+      isControllerBusy = true;
+      lastShootingButtonPressedPosition = null;
     });
     try {
       final XFile file = await controller.stopVideoRecording();
@@ -982,7 +994,7 @@ class CameraPickerState extends State<CameraPicker>
         pickerConfig.onMinimumRecordDurationNotMet?.call();
         return;
       }
-      await controller.pausePreview();
+      controller.pausePreview();
       final bool? isCapturedFileHandled = pickerConfig.onXFileCaptured?.call(
         file,
         CameraPickerViewType.video,
@@ -1000,12 +1012,14 @@ class CameraPickerState extends State<CameraPicker>
         await controller.resumePreview();
       }
     } catch (e, s) {
-      handleError();
+      recordCountdownTimer?.cancel();
       initCameras();
       handleErrorWithHandler(e, s, pickerConfig.onError);
     } finally {
-      isControllerBusy = false;
-      safeSetState(() {});
+      safeSetState(() {
+        isControllerBusy = false;
+        isShootingButtonAnimate = false;
+      });
     }
   }
 
@@ -1319,17 +1333,17 @@ class CameraPickerState extends State<CameraPicker>
                       ),
                     ),
                   ),
-                  if ((innerController?.value.isRecordingVideo ?? false) &&
-                      isRecordingRestricted)
+                  if (shouldCaptureButtonDisplay)
                     RotatedBox(
                       quarterTurns:
                           !enableScaledPreview ? cameraQuarterTurns : 0,
                       child: CameraProgressButton(
                         isAnimating: isShootingButtonAnimate,
+                        isBusy: isControllerBusy,
                         duration: pickerConfig.maximumRecordingDuration!,
-                        outerRadius: outerSize.width,
+                        size: outerSize,
                         ringsColor: theme.indicatorColor,
-                        ringsWidth: 2,
+                        ringsWidth: 3,
                       ),
                     ),
                 ],
@@ -1675,7 +1689,7 @@ class CameraPickerState extends State<CameraPicker>
     // Scale the preview if the config is enabled.
     if (enableScaledPreview) {
       preview = Transform.scale(
-        scale: effectiveCameraScale(constraints, controller),
+        scale: effectiveCameraScale(constraints, innerController),
         child: Center(child: transformedWidget ?? preview),
       );
       // Rotated the preview if the turns is valid.