+++ lib/src/widgets/camera_progress_button.dart
@@ -2,8 +2,6 @@
 // Use of this source code is governed by an Apache license that can be found
 // in the LICENSE file.
 
-import 'dart:math' as math;
-
 import 'package:flutter/material.dart';
 
 import '../constants/styles.dart';
@@ -13,18 +11,18 @@ class CameraProgressButton extends StatefulWidget {
   const CameraProgressButton({
     super.key,
     required this.isAnimating,
-    required this.outerRadius,
+    required this.isBusy,
+    required this.size,
     required this.ringsWidth,
     this.ringsColor = wechatThemeColor,
-    this.progress = 0.0,
     this.duration = const Duration(seconds: 15),
   });
 
   final bool isAnimating;
-  final double outerRadius;
+  final bool isBusy;
+  final Size size;
   final double ringsWidth;
   final Color ringsColor;
-  final double progress;
   final Duration duration;
 
   @override
@@ -33,16 +31,15 @@ class CameraProgressButton extends StatefulWidget {
 
 class _CircleProgressState extends State<CameraProgressButton>
     with SingleTickerProviderStateMixin {
-  final GlobalKey paintKey = GlobalKey();
-
-  late final AnimationController progressController = AnimationController(
-    duration: widget.duration,
-    vsync: this,
-  )..value = widget.progress;
+  late final AnimationController progressController;
 
   @override
   void initState() {
     super.initState();
+    progressController = AnimationController(
+      duration: widget.duration,
+      vsync: this,
+    );
     ambiguate(WidgetsBinding.instance)?.addPostFrameCallback((_) {
       if (widget.isAnimating) {
         progressController.forward();
@@ -53,6 +50,18 @@ class _CircleProgressState extends State<CameraProgressButton>
   @override
   void didUpdateWidget(CameraProgressButton oldWidget) {
     super.didUpdateWidget(oldWidget);
+    if (widget.isBusy != oldWidget.isBusy) {
+      if (widget.isBusy) {
+        progressController
+          ..reset()
+          ..stop();
+      } else {
+        progressController.value = 0.0;
+        if (!progressController.isAnimating) {
+          progressController.forward();
+        }
+      }
+    }
     if (widget.isAnimating != oldWidget.isAnimating) {
       if (widget.isAnimating) {
         progressController.forward();
@@ -70,67 +79,23 @@ class _CircleProgressState extends State<CameraProgressButton>
 
   @override
   Widget build(BuildContext context) {
-    final Size size = Size.square(widget.outerRadius * 2);
+    if (!widget.isAnimating && !widget.isBusy) {
+      return const SizedBox.shrink();
+    }
     return Center(
-      child: RepaintBoundary(
-        child: AnimatedBuilder(
-          animation: progressController,
-          builder: (_, __) => CustomPaint(
-            key: paintKey,
-            size: size,
-            painter: CameraProgressButtonPainter(
-              progress: progressController.value,
-              ringsWidth: widget.ringsWidth,
-              ringsColor: widget.ringsColor,
+      child: SizedBox.fromSize(
+        size: widget.size,
+        child: RepaintBoundary(
+          child: AnimatedBuilder(
+            animation: progressController,
+            builder: (_, __) => CircularProgressIndicator(
+              color: widget.ringsColor,
+              strokeWidth: widget.ringsWidth,
+              value: widget.isBusy ? null : progressController.value,
             ),
           ),
         ),
       ),
     );
-  }
-}
-
-class CameraProgressButtonPainter extends CustomPainter {
-  const CameraProgressButtonPainter({
-    required this.ringsWidth,
-    required this.ringsColor,
-    required this.progress,
-  });
-
-  final double ringsWidth;
-  final Color ringsColor;
-  final double progress;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final double center = size.width / 2;
-    final Offset offsetCenter = Offset(center, center);
-    final double drawRadius = size.width / 2 - ringsWidth;
-
-    final double outerRadius = center;
-    final double innerRadius = center - ringsWidth * 2;
-
-    final double progressWidth = outerRadius - innerRadius;
-    canvas.save();
-    canvas.translate(0.0, size.width);
-    canvas.rotate(-math.pi / 2);
-    final Rect arcRect = Rect.fromCircle(
-      center: offsetCenter,
-      radius: drawRadius,
-    );
-    final Paint progressPaint = Paint()
-      ..color = ringsColor
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = progressWidth;
-    canvas
-      ..drawArc(arcRect, 0, math.pi * 2 * progress, false, progressPaint)
-      ..restore();
-  }
-
-  @override
-  bool shouldRepaint(CameraProgressButtonPainter oldDelegate) {
-    return oldDelegate.ringsWidth != ringsWidth ||
-        oldDelegate.ringsColor != ringsColor ||
-        oldDelegate.progress != progress;
   }
 }